# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task & Payment API",
        "description": "API documentation for Task and Payment endpoints.",
        "version": "1.0.0"
    },
    "paths": {
        "/tasks": {
            "get": {
                "summary": "Get all tasks",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "List of tasks"
                    }
                }
            },
            "post": {
                "summary": "Create a new task",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "summary": "Get a particular task by ID",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details"
                    }
                }
            },
            "delete": {
                "summary": "Delete a task by ID",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully"
                    }
                }
            }
        }
    }
}